Fixed for gcc 4.7

+++ Source/Drivers/PS1080/PS1080Console/PS1080Console.cpp
@@ -953,7 +953,7 @@ bool Download(openni::Device& Device, vector<string>& Command)
 		printf("Can't understand %s as id\n", Command[1].c_str());
 		return true;
 	}
-	XnParamFileData ParamFile = {nId, Command[2].c_str()};
+	XnParamFileData ParamFile = { (uint32_t) nId, Command[2].c_str()};
 
 	openni::Status rc = Device.getProperty(XN_MODULE_PROPERTY_FILE, &ParamFile);
 	if (rc != openni::STATUS_OK)
